feat: enhance context management with modular architecture and improved file handling
- Introduced a new modular structure for context management, separating concerns into distinct classes: `ContextTracker`, `FileCollector`, `FileFilter`, `OutputFormatter`, and `ConfigurationManager`.
- Improved file handling by implementing a dedicated `FileCollector` to manage open files and a `FileFilter` to validate documents against ignore patterns.
- Enhanced error handling with a centralized `ErrorHandler` class for better error reporting.
- Updated the configuration management to support dynamic loading and validation of output settings.
- Refactored the main extension logic to utilize the new modular components, improving maintainability and readability.
