+++ .gitignore
@@ -1,12 +1,9 @@
 out
 dist
-.cursor/rules/context-output.mdc
-.cursor/rules/example.mdc
 output
 node_modules
 .vscode-test/
 *.vsix
-prompt.md
 
 # Added by Task Master AI
 # Logs
