+++ .roo/mcp.json
@@ -2,7 +2,45 @@
   "mcpServers": {
     "context7": {
       "command": "npx",
-      "args": ["-y", "@upstash/context7-mcp@latest"]
+      "args": [
+        "-y",
+        "@upstash/context7-mcp@latest"
+      ]
+    },
+    "task-master-ai": {
+      "command": "npx",
+      "args": [
+        "-y",
+        "--package=task-master-ai",
+        "task-master-ai"
+      ],
+      "env": {},
+      "alwaysAllow": [
+        "initialize_project",
+        "models",
+        "parse_prd",
+        "get_tasks",
+        "get_task",
+        "next_task",
+        "complexity_report",
+        "set_task_status",
+        "generate",
+        "add_task",
+        "add_subtask",
+        "update",
+        "update_task",
+        "update_subtask",
+        "remove_task",
+        "clear_subtasks",
+        "move_task",
+        "analyze_project_complexity",
+        "expand_task",
+        "expand_all",
+        "add_dependency",
+        "remove_dependency",
+        "validate_dependencies",
+        "fix_dependencies"
+      ]
     }
   }
 }
