+++ src/services/OutputFormatter.ts
@@ -1,0 +1,11 @@
+import { FileData, IOutputFormatter } from '../types/interfaces';
+
+export class OutputFormatter implements IOutputFormatter {
+	public formatOutput(files: FileData[], format: string): string {
+		return files.map(file => 
+			format
+				.replace('${fileName}', file.filePath)
+				.replace('${content}', file.content)
+		).join('');
+	}
+}
